+++ boards.go
@@ -24,15 +24,15 @@ import (
 type (
 	IssueBoardsServiceInterface interface {
 		CreateIssueBoard(pid any, opt *CreateIssueBoardOptions, options ...RequestOptionFunc) (*IssueBoard, *Response, error)
-		UpdateIssueBoard(pid any, board int, opt *UpdateIssueBoardOptions, options ...RequestOptionFunc) (*IssueBoard, *Response, error)
-		DeleteIssueBoard(pid any, board int, options ...RequestOptionFunc) (*Response, error)
+		UpdateIssueBoard(pid any, board int64, opt *UpdateIssueBoardOptions, options ...RequestOptionFunc) (*IssueBoard, *Response, error)
+		DeleteIssueBoard(pid any, board int64, options ...RequestOptionFunc) (*Response, error)
 		ListIssueBoards(pid any, opt *ListIssueBoardsOptions, options ...RequestOptionFunc) ([]*IssueBoard, *Response, error)
-		GetIssueBoard(pid any, board int, options ...RequestOptionFunc) (*IssueBoard, *Response, error)
-		GetIssueBoardLists(pid any, board int, opt *GetIssueBoardListsOptions, options ...RequestOptionFunc) ([]*BoardList, *Response, error)
-		GetIssueBoardList(pid any, board, list int, options ...RequestOptionFunc) (*BoardList, *Response, error)
-		CreateIssueBoardList(pid any, board int, opt *CreateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error)
-		UpdateIssueBoardList(pid any, board, list int, opt *UpdateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error)
-		DeleteIssueBoardList(pid any, board, list int, options ...RequestOptionFunc) (*Response, error)
+		GetIssueBoard(pid any, board int64, options ...RequestOptionFunc) (*IssueBoard, *Response, error)
+		GetIssueBoardLists(pid any, board int64, opt *GetIssueBoardListsOptions, options ...RequestOptionFunc) ([]*BoardList, *Response, error)
+		GetIssueBoardList(pid any, board, list int64, options ...RequestOptionFunc) (*BoardList, *Response, error)
+		CreateIssueBoardList(pid any, board int64, opt *CreateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error)
+		UpdateIssueBoardList(pid any, board, list int64, opt *UpdateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error)
+		DeleteIssueBoardList(pid any, board, list int64, options ...RequestOptionFunc) (*Response, error)
 	}
 
 	// IssueBoardsService handles communication with the issue board related
@@ -50,13 +50,13 @@ var _ IssueBoardsServiceInterface = (*IssueBoardsService)(nil)
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/
 type IssueBoard struct {
-	ID        int             `json:"id"`
+	ID        int64           `json:"id"`
 	Name      string          `json:"name"`
 	Project   *Project        `json:"project"`
 	Milestone *Milestone      `json:"milestone"`
 	Assignee  *BasicUser      `json:"assignee"`
 	Lists     []*BoardList    `json:"lists"`
-	Weight    int             `json:"weight"`
+	Weight    int64           `json:"weight"`
 	Labels    []*LabelDetails `json:"labels"`
 }
 
@@ -68,14 +68,14 @@ func (b IssueBoard) String() string {
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/
 type BoardList struct {
-	ID             int                `json:"id"`
+	ID             int64              `json:"id"`
 	Assignee       *BoardListAssignee `json:"assignee"`
 	Iteration      *ProjectIteration  `json:"iteration"`
 	Label          *Label             `json:"label"`
-	MaxIssueCount  int                `json:"max_issue_count"`
-	MaxIssueWeight int                `json:"max_issue_weight"`
+	MaxIssueCount  int64              `json:"max_issue_count"`
+	MaxIssueWeight int64              `json:"max_issue_weight"`
 	Milestone      *Milestone         `json:"milestone"`
-	Position       int                `json:"position"`
+	Position       int64              `json:"position"`
 }
 
 func (b BoardList) String() string {
@@ -86,7 +86,7 @@ func (b BoardList) String() string {
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/
 type BoardListAssignee struct {
-	ID       int    `json:"id"`
+	ID       int64  `json:"id"`
 	Name     string `json:"name"`
 	Username string `json:"username"`
 }
@@ -131,16 +131,16 @@ func (s *IssueBoardsService) CreateIssueBoard(pid any, opt *CreateIssueBoardOpti
 // GitLab API docs: https://docs.gitlab.com/api/boards/#update-an-issue-board
 type UpdateIssueBoardOptions struct {
 	Name        *string       `url:"name,omitempty" json:"name,omitempty"`
-	AssigneeID  *int          `url:"assignee_id,omitempty" json:"assignee_id,omitempty"`
-	MilestoneID *int          `url:"milestone_id,omitempty" json:"milestone_id,omitempty"`
+	AssigneeID  *int64        `url:"assignee_id,omitempty" json:"assignee_id,omitempty"`
+	MilestoneID *int64        `url:"milestone_id,omitempty" json:"milestone_id,omitempty"`
 	Labels      *LabelOptions `url:"labels,omitempty" json:"labels,omitempty"`
-	Weight      *int          `url:"weight,omitempty" json:"weight,omitempty"`
+	Weight      *int64        `url:"weight,omitempty" json:"weight,omitempty"`
 }
 
 // UpdateIssueBoard update an issue board.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#update-an-issue-board
-func (s *IssueBoardsService) UpdateIssueBoard(pid any, board int, opt *UpdateIssueBoardOptions, options ...RequestOptionFunc) (*IssueBoard, *Response, error) {
+func (s *IssueBoardsService) UpdateIssueBoard(pid any, board int64, opt *UpdateIssueBoardOptions, options ...RequestOptionFunc) (*IssueBoard, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -164,7 +164,7 @@ func (s *IssueBoardsService) UpdateIssueBoard(pid any, board int, opt *UpdateIss
 // DeleteIssueBoard deletes an issue board.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#delete-an-issue-board
-func (s *IssueBoardsService) DeleteIssueBoard(pid any, board int, options ...RequestOptionFunc) (*Response, error) {
+func (s *IssueBoardsService) DeleteIssueBoard(pid any, board int64, options ...RequestOptionFunc) (*Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, err
@@ -213,7 +213,7 @@ func (s *IssueBoardsService) ListIssueBoards(pid any, opt *ListIssueBoardsOption
 // GetIssueBoard gets a single issue board of a project.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#show-a-single-issue-board
-func (s *IssueBoardsService) GetIssueBoard(pid any, board int, options ...RequestOptionFunc) (*IssueBoard, *Response, error) {
+func (s *IssueBoardsService) GetIssueBoard(pid any, board int64, options ...RequestOptionFunc) (*IssueBoard, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -245,7 +245,7 @@ type GetIssueBoardListsOptions struct {
 // backlog and closed lists.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#list-board-lists-in-a-project-issue-board
-func (s *IssueBoardsService) GetIssueBoardLists(pid any, board int, opt *GetIssueBoardListsOptions, options ...RequestOptionFunc) ([]*BoardList, *Response, error) {
+func (s *IssueBoardsService) GetIssueBoardLists(pid any, board int64, opt *GetIssueBoardListsOptions, options ...RequestOptionFunc) ([]*BoardList, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -269,7 +269,7 @@ func (s *IssueBoardsService) GetIssueBoardLists(pid any, board int, opt *GetIssu
 // GetIssueBoardList gets a single issue board list.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#show-a-single-board-list
-func (s *IssueBoardsService) GetIssueBoardList(pid any, board, list int, options ...RequestOptionFunc) (*BoardList, *Response, error) {
+func (s *IssueBoardsService) GetIssueBoardList(pid any, board, list int64, options ...RequestOptionFunc) (*BoardList, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -299,16 +299,16 @@ func (s *IssueBoardsService) GetIssueBoardList(pid any, board, list int, options
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#create-a-board-list
 type CreateIssueBoardListOptions struct {
-	LabelID     *int `url:"label_id,omitempty" json:"label_id,omitempty"`
-	AssigneeID  *int `url:"assignee_id,omitempty" json:"assignee_id,omitempty"`
-	MilestoneID *int `url:"milestone_id,omitempty" json:"milestone_id,omitempty"`
-	IterationID *int `url:"iteration_id,omitempty" json:"iteration_id,omitempty"`
+	LabelID     *int64 `url:"label_id,omitempty" json:"label_id,omitempty"`
+	AssigneeID  *int64 `url:"assignee_id,omitempty" json:"assignee_id,omitempty"`
+	MilestoneID *int64 `url:"milestone_id,omitempty" json:"milestone_id,omitempty"`
+	IterationID *int64 `url:"iteration_id,omitempty" json:"iteration_id,omitempty"`
 }
 
 // CreateIssueBoardList creates a new issue board list.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#create-a-board-list
-func (s *IssueBoardsService) CreateIssueBoardList(pid any, board int, opt *CreateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error) {
+func (s *IssueBoardsService) CreateIssueBoardList(pid any, board int64, opt *CreateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -334,13 +334,13 @@ func (s *IssueBoardsService) CreateIssueBoardList(pid any, board int, opt *Creat
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#reorder-a-list-in-a-board
 type UpdateIssueBoardListOptions struct {
-	Position *int `url:"position" json:"position"`
+	Position *int64 `url:"position" json:"position"`
 }
 
 // UpdateIssueBoardList updates the position of an existing issue board list.
 //
 // GitLab API docs: https://docs.gitlab.com/api/boards/#reorder-a-list-in-a-board
-func (s *IssueBoardsService) UpdateIssueBoardList(pid any, board, list int, opt *UpdateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error) {
+func (s *IssueBoardsService) UpdateIssueBoardList(pid any, board, list int64, opt *UpdateIssueBoardListOptions, options ...RequestOptionFunc) (*BoardList, *Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, nil, err
@@ -370,7 +370,7 @@ func (s *IssueBoardsService) UpdateIssueBoardList(pid any, board, list int, opt 
 //
 // GitLab API docs:
 // https://docs.gitlab.com/api/boards/#delete-a-board-list-from-a-board
-func (s *IssueBoardsService) DeleteIssueBoardList(pid any, board, list int, options ...RequestOptionFunc) (*Response, error) {
+func (s *IssueBoardsService) DeleteIssueBoardList(pid any, board, list int64, options ...RequestOptionFunc) (*Response, error) {
 	project, err := parseID(pid)
 	if err != nil {
 		return nil, err
